requirements.txt
sentence-transformers
faiss-cpu
PyMuPDF
ipykernel
ollama
streamlit